refactor(svelte): split inspection-state test monolith into harness suites
Move createInspectionState composite tests out of a single ~1k-line file into
a shared harness plus three responsibility-based suites (construction/effects,
pin/queue, dismiss/traversal). No production API change — pure decision tables
already live under inspection/; keep the imperative shell co-located.

Closes #431
