+++ packages/svelte/tests/inspection/inspection-state.harness.ts
@@ -1,0 +1,221 @@
+/**
+ * Shared harness for createInspectionState composite tests.
+ * Factories own deriveds + effects — instantiate under `$effect.root` and destroy.
+ */
+import type { CandidateFacts, CellValue, RenderModel } from "@ggsvelte/core";
+import { aes, gg, type PortableSpec } from "@ggsvelte/spec";
+
+import type {
+  PlotInspection,
+  PlotInteractionEvent,
+  ResolvedInteractionConfig,
+} from "../../src/lib/interaction/interaction.js";
+import { normalizeInteractionConfig } from "../../src/lib/interaction/interaction.js";
+import { createInteractionReducer } from "../../src/lib/interaction/reducer.js";
+import {
+  createInspectionState,
+  type InspectionStateDeps,
+} from "../../src/lib/inspection/inspection-state.svelte.js";
+import { hitFromCandidate } from "../../src/lib/surface/plot-px.js";
+import { withFlushedEffectRoot } from "../helpers/effect-root.svelte.js";
+import { modelFor } from "../helpers/model.js";
+
+const continuousRows = [
+  { id: "a", x: 1, y: 1 },
+  { id: "b", x: 10, y: 20 },
+  { id: "c", x: 5, y: 8 },
+];
+
+export type InspectConfig = ResolvedInteractionConfig["inspect"];
+export type InteractionReducer = ReturnType<typeof createInteractionReducer>;
+export type InspectCb = ((event: PlotInspection<Record<string, CellValue>>) => void) | undefined;
+export type InteractionCb =
+  | ((event: PlotInteractionEvent<Record<string, CellValue>>) => void)
+  | undefined;
+
+export const noInspect: () => InspectCb = () => {
+  /* no callback */
+};
+export const noInteraction: () => InteractionCb = () => {
+  /* no callback */
+};
+
+export const defaultInspect = (): InspectConfig =>
+  normalizeInteractionConfig({ inspect: { pin: true } }).inspect;
+
+export const modeXInspect = (): InspectConfig =>
+  normalizeInteractionConfig({ inspect: { pin: false, mode: "x" } }).inspect;
+
+/** Identity semantic keys — enough for local consumption without a host service. */
+export function keyAtForModel(model: RenderModel): (index: number) => PropertyKey | null {
+  return (index) => {
+    const row = model.row(index);
+    if (row === null) return null;
+    const id = row["id"];
+    return id === undefined || id === null ? String(index) : String(id);
+  };
+}
+
+export function continuousSpec(
+  data: readonly { id: string; x: number; y: number }[] = continuousRows,
+): PortableSpec {
+  return gg([...data], aes({ x: "x", y: "y" }))
+    .geomPoint()
+    .spec();
+}
+
+/**
+ * Twelve series at the same x (color-split points) so mode "x" groups more
+ * than the transient eight-member limit (matches interaction-unit fixture).
+ */
+export function largeGroupSpec(): PortableSpec {
+  const rows = Array.from({ length: 12 }, (_, i) => ({
+    id: `row-${i}`,
+    x: 1,
+    y: i,
+    series: `series-${i}`,
+  }));
+  return gg(rows, aes({ x: "x", y: "y", color: "series" }))
+    .geomPoint()
+    .spec();
+}
+
+export function firstCandidate(model: RenderModel): CandidateFacts {
+  for (let id = 0; id < model.candidates.size; id++) {
+    const candidate = model.candidates.candidate(id);
+    if (candidate !== null) return candidate;
+  }
+  throw new Error("expected at least one candidate");
+}
+
+export function candidateHit(model: RenderModel): {
+  candidate: CandidateFacts;
+  hit: ReturnType<typeof hitFromCandidate>;
+} {
+  const candidate = firstCandidate(model);
+  return { candidate, hit: hitFromCandidate(candidate) };
+}
+
+type ConcreteMode = "exact" | "x" | "y" | "xy";
+
+/** Apply transient/pinned inspection without threading CandidateFacts.mode (oxlint). */
+export function applyInspect(
+  state: ReturnType<typeof createInspectionState>,
+  candidate: CandidateFacts,
+  source: "pointer" | "keyboard" | "touch" = "pointer",
+  inspectionState: "transient" | "pinned" = "transient",
+  concreteMode: ConcreteMode = "xy",
+): void {
+  state.setInspection(
+    hitFromCandidate(candidate),
+    source,
+    inspectionState,
+    concreteMode,
+    candidate,
+  );
+}
+
+/** Wrap a reducer so inspect-null dispatches are logged (shared spy). */
+export function wrapReducerWithLog(
+  base: InteractionReducer,
+  onInspectNull: (count: number) => void,
+): InteractionReducer {
+  let nullDispatches = 0;
+  return {
+    ...base,
+    dispatch: (action) => {
+      if (action.type === "inspect" && action.candidate === null) {
+        nullDispatches++;
+        onInspectNull(nullDispatches);
+      }
+      base.dispatch(action);
+    },
+    get state() {
+      return base.state;
+    },
+  };
+}
+
+export type InspectionHarness = {
+  state: ReturnType<typeof createInspectionState>;
+  reducer: InteractionReducer;
+  destroy: () => void;
+};
+
+/**
+ * Mount the controller with production-shaped deps: every reactive dep is a
+ * getter (mirroring InspectionStateDeps). Harness owns the component-held
+ * reducer and passes a getter. Call registerInspectionEffects when effects
+ * are under test.
+ */
+export function mountInspectionController(
+  options: {
+    model?: () => RenderModel | null;
+    reducer?: () => InteractionReducer;
+    inspectConfig?: () => InspectConfig;
+    inspectEnabled?: () => boolean;
+    dataIdentityEpoch?: () => string;
+    keyAt?: (index: number) => PropertyKey | null;
+    root?: () => HTMLDivElement | null;
+    captureSurface?: () => HTMLDivElement | null;
+    plotId?: () => string;
+    tooltipHovered?: () => boolean;
+    clearTooltipHovered?: () => void;
+    clearBrush?: () => void;
+    chooseTool?: (tool: "inspect") => void;
+    oninspect?: InspectionStateDeps["oninspect"];
+    oninteraction?: InspectionStateDeps["oninteraction"];
+    announce?: (message: string) => void;
+    clearAnnouncement?: () => void;
+    registerEffects?: boolean;
+  } = {},
+): InspectionHarness {
+  const defaultModel = modelFor(continuousSpec());
+  const ownedReducer =
+    options.reducer === undefined
+      ? createInteractionReducer({
+          scheduleFrame: (callback) => {
+            callback();
+            return 0;
+          },
+          cancelFrame: () => {},
+        })
+      : null;
+  const getReducer = options.reducer ?? (() => ownedReducer!);
+
+  const { value: state, destroy } = withFlushedEffectRoot(() => {
+    const controller = createInspectionState({
+      model: options.model ?? (() => defaultModel),
+      reducer: getReducer,
+      inspectConfig: options.inspectConfig ?? defaultInspect,
+      inspectEnabled: options.inspectEnabled ?? (() => true),
+      dataIdentityEpoch: options.dataIdentityEpoch ?? (() => "epoch-1"),
+      keyAt:
+        options.keyAt ??
+        ((index) => {
+          const model = options.model?.() ?? defaultModel;
+          return keyAtForModel(model)(index);
+        }),
+      root: options.root ?? (() => null),
+      captureSurface: options.captureSurface ?? (() => null),
+      plotId: options.plotId ?? (() => "plot-test"),
+      tooltipHovered: options.tooltipHovered ?? (() => false),
+      clearTooltipHovered: options.clearTooltipHovered ?? (() => {}),
+      clearBrush: options.clearBrush ?? (() => {}),
+      chooseTool: options.chooseTool ?? (() => {}),
+      oninspect: options.oninspect ?? noInspect,
+      oninteraction: options.oninteraction ?? noInteraction,
+      announce: options.announce ?? (() => {}),
+      clearAnnouncement: options.clearAnnouncement ?? (() => {}),
+    });
+    if (options.registerEffects !== false) controller.registerInspectionEffects();
+    return controller;
+  });
+
+  return { state, reducer: getReducer(), destroy };
+}
+
+// Re-exports used by construction / armed-getter suites
+export { createInspectionState, createInteractionReducer, modelFor, hitFromCandidate };
+export { withFlushedEffectRoot } from "../helpers/effect-root.svelte.js";
+export type { InspectionStateDeps, RenderModel, CandidateFacts, PortableSpec };
